Module Tezos_client_003_PsddFKi3.Michelson_v1_parser

type parsed = {

source : string;
The original source code.

unexpanded : string Tezos_micheline.Micheline.canonical;
Original expression with macros.

expanded : Tezos_raw_protocol_003_PsddFKi3.Alpha_context.Script.expr;
Expression with macros fully expanded.

expansion_table : (int * (Tezos_micheline.Micheline_parser.location * int list)) list;
Associates unexpanded nodes to their parsing locations and the nodes expanded from it in the expanded expression.

unexpansion_table : (int * int) list;
Associates an expanded node to its source in the unexpanded expression.

}

The result of parsing and expanding a Michelson V1 script or data.

val compare_parsed : parsed -> parsed -> int

val parse_toplevel : ?⁠check:bool -> string -> parsed Tezos_micheline.Micheline_parser.parsing_result

val parse_expression : ?⁠check:bool -> string -> parsed Tezos_micheline.Micheline_parser.parsing_result

val expand_all : source:string -> original:Tezos_micheline.Micheline_parser.node -> parsed Tezos_micheline.Micheline_parser.parsing_result